refactor: split checkpoint management modules
