+++ crates/bitnet-models/src/checkpoint.rs
@@ -4,342 +4,23 @@
 //! verification (SHA-256), and a thread-safe inventory for managing model
 //! checkpoint files across supported formats (GGUF, SafeTensors, PyTorch,
 //! Custom).
+//!
+//! The implementation is split by single responsibility:
+//!
+//! - `format` detects supported checkpoint encodings.
+//! - `metadata` extracts filesystem metadata and integrity hashes.
+//! - `manager` owns the thread-safe checkpoint inventory.
+//! - `error` defines the shared error type.
 
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
-use std::collections::HashMap;
-use std::io::Read;
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, RwLock};
-use std::time::SystemTime;
+mod error;
+mod format;
+mod manager;
+mod metadata;
 
-// ---------------------------------------------------------------------------
-// CheckpointFormat
-// ---------------------------------------------------------------------------
-
-/// Supported model checkpoint formats.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum CheckpointFormat {
-    /// GGUF (llama.cpp / ggml ecosystem).
-    Gguf,
-    /// SafeTensors (HuggingFace standard).
-    SafeTensors,
-    /// PyTorch serialised checkpoint (`.pt` / `.bin` / `.pth`).
-    PyTorch,
-    /// User-defined / unrecognised format.
-    Custom,
-}
-
-impl CheckpointFormat {
-    /// Detect checkpoint format from a file path using extension heuristics
-    /// followed by a header probe.
-    pub fn detect(path: &Path) -> Self {
-        if let Some(ext) = path.extension().and_then(|e| e.to_str()) {
-            match ext.to_lowercase().as_str() {
-                "gguf" => return Self::Gguf,
-                "safetensors" => return Self::SafeTensors,
-                "pt" | "pth" | "bin" => return Self::PyTorch,
-                _ => {}
-            }
-        }
-        // Fall back to header magic bytes when extension is absent or
-        // unrecognised.
-        Self::detect_from_header(path).unwrap_or(Self::Custom)
-    }
-
-    /// Inspect the first bytes of `path` for known magic values.
-    fn detect_from_header(path: &Path) -> Option<Self> {
-        let mut file = std::fs::File::open(path).ok()?;
-        let mut header = [0u8; 8];
-        file.read_exact(&mut header).ok()?;
-
-        // GGUF v3 magic: "GGUF" as LE u32 = 0x46475547
-        if header[..4] == [0x47, 0x47, 0x55, 0x46] {
-            return Some(Self::Gguf);
-        }
-        // SafeTensors files start with a little-endian u64 length followed by
-        // JSON — the first byte is usually a small number (< 256) while bytes
-        // 4..8 are zero for any header < 4 GiB, and the JSON typically starts
-        // with '{'. We use a lightweight heuristic: if byte-8 would be '{'
-        // (0x7b) we treat it as SafeTensors.
-        if header[0] > 0 && header[4..8] == [0, 0, 0, 0] {
-            // Read one more byte to check for JSON open brace.
-            let mut json_byte = [0u8; 1];
-            if file.read_exact(&mut json_byte).is_ok() && json_byte[0] == b'{' {
-                return Some(Self::SafeTensors);
-            }
-        }
-        // PyTorch ZIP-based checkpoints start with the PK magic.
-        if header[..2] == [0x50, 0x4B] {
-            return Some(Self::PyTorch);
-        }
-        None
-    }
-
-    /// Human-readable label.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Gguf => "GGUF",
-            Self::SafeTensors => "SafeTensors",
-            Self::PyTorch => "PyTorch",
-            Self::Custom => "Custom",
-        }
-    }
-}
-
-impl std::fmt::Display for CheckpointFormat {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-// ---------------------------------------------------------------------------
-// CheckpointMetadata
-// ---------------------------------------------------------------------------
-
-/// Metadata associated with a single model checkpoint file.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct CheckpointMetadata {
-    /// Detected or overridden checkpoint format.
-    pub format: CheckpointFormat,
-    /// Model name (derived from the file stem by default).
-    pub model_name: String,
-    /// Optional version string.
-    pub version: Option<String>,
-    /// Timestamp when the metadata entry was created.
-    pub created_at: SystemTime,
-    /// File size in bytes.
-    pub file_size: u64,
-    /// SHA-256 hex digest of the file contents.
-    pub hash: String,
-    /// Canonical path to the checkpoint file.
-    pub path: PathBuf,
-    /// Last modification time reported by the filesystem.
-    pub modified_at: Option<SystemTime>,
-}
-
-// ---------------------------------------------------------------------------
-// CheckpointError
-// ---------------------------------------------------------------------------
-
-/// Errors specific to checkpoint management.
-#[derive(Debug, thiserror::Error)]
-pub enum CheckpointError {
-    #[error("checkpoint not found: {0}")]
-    NotFound(String),
-    #[error("duplicate checkpoint: {0}")]
-    Duplicate(String),
-    #[error("hash mismatch for {path}: expected {expected}, got {actual}")]
-    HashMismatch { path: String, expected: String, actual: String },
-    #[error("IO error: {0}")]
-    Io(#[from] std::io::Error),
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/// Compute the SHA-256 hex digest of a file using a streaming 1 MiB buffer.
-pub fn compute_sha256(path: &Path) -> Result<String, CheckpointError> {
-    let mut file = std::fs::File::open(path)?;
-    let mut hasher = Sha256::new();
-    let mut buf = vec![0u8; 1024 * 1024];
-    loop {
-        let n = file.read(&mut buf)?;
-        if n == 0 {
-            break;
-        }
-        hasher.update(&buf[..n]);
-    }
-    Ok(format!("{:x}", hasher.finalize()))
-}
-
-/// Derive a model name from a file path (uses the file stem).
-fn model_name_from_path(path: &Path) -> String {
-    path.file_stem().and_then(|s| s.to_str()).unwrap_or("unknown").to_string()
-}
-
-/// Build a stable inventory key for a path.
-///
-/// We prefer canonicalized paths to deduplicate relative/absolute aliases that
-/// point to the same file, but fall back to an absolute best-effort key when
-/// canonicalization is unavailable (e.g. path does not currently exist).
-fn inventory_key(path: &Path) -> String {
-    let normalized = std::fs::canonicalize(path).unwrap_or_else(|_| {
-        if path.is_absolute() {
-            path.to_path_buf()
-        } else {
-            std::env::current_dir().map(|cwd| cwd.join(path)).unwrap_or_else(|_| path.to_path_buf())
-        }
-    });
-    normalized.to_string_lossy().to_string()
-}
-
-/// Build [`CheckpointMetadata`] by inspecting a file on disk.
-pub fn extract_metadata(path: &Path) -> Result<CheckpointMetadata, CheckpointError> {
-    let meta = std::fs::metadata(path)?;
-    let hash = compute_sha256(path)?;
-    let format = CheckpointFormat::detect(path);
-    let modified_at = meta.modified().ok();
-    let canonical_path = std::fs::canonicalize(path).unwrap_or_else(|_| path.to_path_buf());
-
-    Ok(CheckpointMetadata {
-        format,
-        model_name: model_name_from_path(path),
-        version: None,
-        created_at: SystemTime::now(),
-        file_size: meta.len(),
-        hash,
-        path: canonical_path,
-        modified_at,
-    })
-}
-
-// ---------------------------------------------------------------------------
-// CheckpointManager
-// ---------------------------------------------------------------------------
-
-/// Thread-safe manager for an inventory of model checkpoints.
-///
-/// All public methods acquire the inner lock for the minimum required
-/// duration, so the manager is safe for concurrent access from multiple
-/// threads.
-#[derive(Debug, Clone)]
-pub struct CheckpointManager {
-    /// Keyed by the canonical string representation of the file path.
-    inventory: Arc<RwLock<HashMap<String, CheckpointMetadata>>>,
-}
-
-impl Default for CheckpointManager {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl CheckpointManager {
-    /// Create an empty manager.
-    pub fn new() -> Self {
-        Self { inventory: Arc::new(RwLock::new(HashMap::new())) }
-    }
-
-    // -- mutating operations ------------------------------------------------
-
-    /// Register a checkpoint file. Extracts metadata from the file system and
-    /// stores it in the inventory. Returns an error if the path is already
-    /// registered.
-    pub fn add(&self, path: &Path) -> Result<CheckpointMetadata, CheckpointError> {
-        let key = inventory_key(path);
-        let meta = extract_metadata(path)?;
-
-        let mut inv = self.inventory.write().expect("lock poisoned");
-        if inv.contains_key(&key) {
-            return Err(CheckpointError::Duplicate(key));
-        }
-        inv.insert(key, meta.clone());
-        Ok(meta)
-    }
-
-    /// Remove a checkpoint from the inventory (does **not** delete the file).
-    pub fn remove(&self, path: &Path) -> Result<CheckpointMetadata, CheckpointError> {
-        let key = inventory_key(path);
-        let mut inv = self.inventory.write().expect("lock poisoned");
-        inv.remove(&key).ok_or(CheckpointError::NotFound(key))
-    }
-
-    // -- read-only queries --------------------------------------------------
-
-    /// Return the number of checkpoints in the inventory.
-    pub fn len(&self) -> usize {
-        self.inventory.read().expect("lock poisoned").len()
-    }
-
-    /// Return `true` when the inventory is empty.
-    pub fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-
-    /// Retrieve metadata for a specific path.
-    pub fn get(&self, path: &Path) -> Option<CheckpointMetadata> {
-        let key = inventory_key(path);
-        self.inventory.read().expect("lock poisoned").get(&key).cloned()
-    }
-
-    /// List all registered checkpoints.
-    pub fn list(&self) -> Vec<CheckpointMetadata> {
-        self.inventory.read().expect("lock poisoned").values().cloned().collect()
-    }
-
-    /// Search by model name (case-insensitive substring match).
-    pub fn search_by_name(&self, query: &str) -> Vec<CheckpointMetadata> {
-        let q = query.to_lowercase();
-        self.inventory
-            .read()
-            .expect("lock poisoned")
-            .values()
-            .filter(|m| m.model_name.to_lowercase().contains(&q))
-            .cloned()
-            .collect()
-    }
-
-    /// Filter inventory by checkpoint format.
-    pub fn filter_by_format(&self, format: CheckpointFormat) -> Vec<CheckpointMetadata> {
-        self.inventory
-            .read()
-            .expect("lock poisoned")
-            .values()
-            .filter(|m| m.format == format)
-            .cloned()
-            .collect()
-    }
-
-    /// Verify the integrity of a registered checkpoint by recomputing its
-    /// SHA-256 digest and comparing against the stored hash.
-    pub fn verify(&self, path: &Path) -> Result<bool, CheckpointError> {
-        let meta = self
-            .get(path)
-            .ok_or_else(|| CheckpointError::NotFound(path.to_string_lossy().to_string()))?;
-        let actual = compute_sha256(path)?;
-        if actual != meta.hash {
-            return Err(CheckpointError::HashMismatch {
-                path: path.to_string_lossy().to_string(),
-                expected: meta.hash.clone(),
-                actual,
-            });
-        }
-        Ok(true)
-    }
-
-    /// Scan a directory for checkpoint files and register every one that
-    /// matches a known extension. Returns the number of files added.
-    pub fn scan_directory(&self, dir: &Path) -> Result<usize, CheckpointError> {
-        let entries = std::fs::read_dir(dir)?;
-        let mut count = 0usize;
-        for entry in entries {
-            let entry = entry?;
-            let path = entry.path();
-            if !path.is_file() {
-                continue;
-            }
-            let dominated = path
-                .extension()
-                .and_then(|e| e.to_str())
-                .map(|e| {
-                    matches!(
-                        e.to_lowercase().as_str(),
-                        "gguf" | "safetensors" | "pt" | "pth" | "bin"
-                    )
-                })
-                .unwrap_or(false);
-            if dominated {
-                // Silently skip duplicates during scanning.
-                if self.add(&path).is_ok() {
-                    count += 1;
-                }
-            }
-        }
-        Ok(count)
-    }
-}
+pub use error::CheckpointError;
+pub use format::CheckpointFormat;
+pub use manager::CheckpointManager;
+pub use metadata::{CheckpointMetadata, compute_sha256, extract_metadata};
 
 // ---------------------------------------------------------------------------
 // Tests
@@ -349,6 +30,7 @@ impl CheckpointManager {
 mod tests {
     use super::*;
     use std::io::Write;
+    use std::path::{Path, PathBuf};
     use tempfile::TempDir;
 
     /// Helper: create a temp file with the given name and contents, return its
